requirements.txt
ruff==0.9.3
python-dotenv==1.0.1
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pydantic==2.10.6
seleniumbase==4.22.3
beautifulsoup4==4.12.2
pytz==2024.2
thefuzz==0.22.1